+++ html/js/config/messages.js
@@ -1,8 +1,11 @@
 import { finishingMessage, serverMessage } from '../util/elements.js';
 
+const colorRegex = /\^[\d]/g;
+
 /**
  * @param {string} str
  * @param {Record<string, string>} vars
+ * @returns {string}
  */
 function replaceVariables(str, vars) {
     for (const [k, v] of Object.entries(vars)) {
@@ -13,10 +16,20 @@ function replaceVariables(str, vars) {
 }
 
 /**
+ * @param {string} message
+ * @returns {string}
+ */
+function removeColors(message) {
+    return message.replaceAll(colorRegex, '');
+}
+
+/**
  * @param {NuiHandoverData} handoverData
  */
 export function configMessages({ vars, config }) {
-    serverMessage.innerText = replaceVariables(config.serverMessage, vars);
+    serverMessage.innerText = removeColors(
+        replaceVariables(config.serverMessage, vars),
+    );
 
-    finishingMessage.innerText = config.finishingMessage;
+    finishingMessage.innerText = removeColors(config.finishingMessage);
 }